update-admin

+++ src/admin/admin.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, ConflictException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Admin } from 'src/DB/models/admin.model';
@@ -16,7 +16,7 @@ export class AdminService {
 
     async createAdmin(body: any, req: any) {
         const { username, password } = body
-        if(req.authAdmin.role != 'king') throw new BadRequestException('Unauthorized, You are not the king admin')
+        if(req.authAdmin.role != 'king') throw new UnauthorizedException('Unauthorized, You are not the king admin')
         const admin = await this.adminModel.findOne({ username })
         if(admin) throw new BadRequestException('Admin username is already exists')
         const hashPassword = bcrypt.hashSync(password, +process.env.SALT_ROUNDS_1)
@@ -27,23 +27,23 @@ export class AdminService {
     async login(body: any) {
         const { username, password } = body
         const admin = await this.adminModel.findOne({ username })
-        if(!admin) throw new ConflictException('Invalid login credentials')
+        if(!admin) throw new BadRequestException('Invalid login credentials')
         if(!bcrypt.compareSync(password, admin.password)) throw new BadRequestException('Invalid login credentials')
         const adminToken = this.jwtService.sign({ id: admin._id, name: admin.username, role: admin.role },
             { secret: process.env.JWT_SECRET_LOGIN, expiresIn: "90d" })
         return adminToken
     }
 
-    async getAdmin(params: any) {
-        const admin = await this.adminModel.findById(params.adminId).select(' username role')
-        if(!admin) throw new ConflictException('Admin not found')
+    async getAdmin(req: any) {
+        const admin = await this.adminModel.findById(req.authAdmin.id).select(' username role')
+        if(!admin) throw new NotFoundException('Admin not found')
         return admin
     }
 
     async getAllAdmins(req: any) {
-        if(req.authAdmin.role != 'king') throw new BadRequestException('Unauthorized, You are not the king admin')
+        if(req.authAdmin.role != 'king') throw new UnauthorizedException('Unauthorized, You are not the king admin')
         const admins = await this.adminModel.find().select(' username role')
-        if(!admins.length) throw new ConflictException('No admins found')
+        if(!admins.length) throw new NotFoundException('No admins found')
         return admins
     }
 
@@ -57,8 +57,10 @@ export class AdminService {
         return true
     }
 
-    async deleteMyAccount(req: any) {
-        const admin = await this.adminModel.findById(req.authAdmin.id)
+    async deleteAdmin(req: any, param: any) {
+        if(req.authAdmin.role != 'king') throw new UnauthorizedException('Unauthorized, You are not the king admin')
+        const admin = await this.adminModel.findById(param.adminId)
+        if(!admin) throw new NotFoundException('Admin not found')
         await admin.deleteOne()
         return true
     }
